fix: singular fields should be optional to write
As per [`proto3`'s language guide](https://developers.google.com/protocol-buffers/docs/proto3#specifying_field_rules):

> Message fields can be one of the following:
>  - singular: a well-formed message can have **zero** or one of this field (but not more than one). And this is the default field rule for proto3 syntax.

This means that all `proto3` fields are effectively optional when
writing, so it's relatively easy to implement by accepting a `Partial`
version of the message interface for encoding.

When reading messages singular fields are initialized to their default
values when reading and optional values are not so the plain non-`Partial`
interface can be returned.

This also has the nice side effect of not requring the user to pass
empty lists/maps for `repeated` and `map` fields which never really
felt right.  These values are initted to their empty forms when
reading messages from the wire.

Fixes #42