fix: remove writing default values
With correct singular/optional handling it's not necessary to write
default values.

This was previously implemented to write default values for repeated
message fields where all fields were singular but those messages are
written as zero-length buffers and their fields are set to the
default value during decoding so there's no reason to have their
values in the buffer.