feat: add custom protons options for limiting list/map sizes
Adds the capability to define protons-specific custom options that
control decoding behaviour initially around limiting the sizes of
maps and lists.

```protobuf
// import the options definition - it will work without this but some
// code editor plugins may report an error if the def can't be found
import "protons-rutime/options.proto";

message MessageWithSizeLimitedRepeatedField {
  // define the size limit - here more than 10 repeated items will
  // cause decoding to fail
  repeated string repeatedField = 1 [(protons.limit) = 10];
}
```

The defintion is shipped with the `protons-runtime` module. There is
a [pending PR](https://github.com/protocolbuffers/protobuf/pull/14505)
to reserve the `1186` field ID. This should be merged first and/or
the field ID updated here if it changes due to that PR.

Fixes #113